Fix autowired and move case service: cases.services -> services.cases
